% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_svg.R
\name{morph_svg}
\alias{morph_svg}
\title{Quarto SVG Morph}
\usage{
morph_svg(
  ...,
  folder = "",
  height = "auto",
  img_id = toString(as.numeric(Sys.time()) * 1e+06),
  animation_duration = 0.2,
  animation_curve = "cubic-bezier(0.76, 0, 0.24, 1)",
  click = T,
  new_ids = F
)
}
\arguments{
\item{...}{Filenames: Add two or more filenames incl. paths. Can also be a vector with filenames.}

\item{folder}{Alternative to using single filename location of two or more svg files}

\item{height}{Height of the SVG file. Defaults to "auto" which should correspond to the height the svg document is saved in. Default is pixels, can be any css size attribute.}

\item{img_id}{Optional internal id.}

\item{animation_duration}{Optional duration of animation in seconds. Defaults to .2s.}

\item{animation_curve}{HTML animation curve. Can be "linear", "ease", "ease-in", "ease-out", "ease-in-out", or any cubic-bezier function. Defaults to: cubic-bezier(0.76, 0, 0.24, 1)}

\item{click}{Makes animation clickable, e.g. for bookdown documents, websites, etc. Defaults to TRUE. In Quarto presentations clickability it follows the reveal structure and click can be turned off if undesirable by setting to FALSE.}

\item{new_ids}{Removes existing ids. Not recommended.}
}
\value{
HTML-Text to output the animation
}
\description{
This function creates animations from two or more SVG files to display in Quarto or other R-generated HTML-Documents. It takes folder names or file names as inputs.
}
\examples{
morph_svg(
  "https://raw.githubusercontent.com/klingelhoefer/quartoMorphSvg/master/tests/images/mediation/Slide1.svg",
  "https://raw.githubusercontent.com/klingelhoefer/quartoMorphSvg/master/tests/images/mediation/Slide2.svg"
  )
}
